+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.1.0] - 2023-07-21
+
+### Added
+- [samber/oops](https://github.com/samber/oops) for better stack trace.
+
+### Changed
+- `workingDir`, `branchFrom`, `branchTo`, and `projects` now defaults to empty string.
+- Throw error if parse flag failed.
+- Removing `sync.WaitGroup{}`
+
+### Fixed
+
+- Fixed flow error.
+- Fixed GitHub Actions version not found.
+
 ## [1.0.1] - 2023-07-03
 
 ### Fixed
@@ -51,3 +66,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed test build.
 
+[1.1.0]: https://github.com/shadowbane/branch-changer/compare/v1.0.1...v1.1.0
+[1.0.1]: https://github.com/shadowbane/branch-changer/compare/v1.0.0...v1.0.1
+[1.0.0]: https://github.com/shadowbane/branch-changer/compare/v0.1.6...v1.0.0
+[0.1.6]: https://github.com/shadowbane/branch-changer/compare/v0.1.5...v0.1.6
+[0.1.5]: https://github.com/shadowbane/branch-changer/compare/v0.1.4...v0.1.5
+[0.1.4]: https://github.com/shadowbane/branch-changer/compare/v0.1.3...v0.1.4
+[0.1.3]: https://github.com/shadowbane/branch-changer/compare/v0.1.2...v0.1.3
+[0.1.2]: https://github.com/shadowbane/branch-changer/releases/tag/v0.1.2